ai-service/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
langchain-google-genai
langchain
python-dotenv
